+++ src/ONTraC/bin/GP.py
@@ -6,12 +6,13 @@ import sys
 import numpy as np
 import torch
 
-from ONTraC.model import GraphPooling
-from ONTraC.optparser import opt_GP_validate, prepare_GP_optparser
-from ONTraC.run.processes import *
-from ONTraC.train import GPBatchTrain, SubBatchTrainProtocol
-from ONTraC.train.inspect_funcs import loss_record
-from ONTraC.utils import device_validate
+from ..model import GraphPooling
+from ..optparser import opt_GP_validate, prepare_GP_optparser
+from ..run.processes import *
+from ..train import GPBatchTrain, SubBatchTrainProtocol
+from ..train.inspect_funcs import loss_record
+from ..utils import device_validate, write_version_info
+from ..utils.niche_net_constr import load_original_data
 
 # ------------------------------------
 # Classes
@@ -40,6 +41,9 @@ def main() -> None:
     :return: None
     """
 
+    # write version information
+    write_version_info()
+    
     # ----- prepare -----
     # load parameters
     options = load_parameters(opt_validate_func=opt_GP_validate, prepare_optparser_func=prepare_GP_optparser)
@@ -73,7 +77,17 @@ def main() -> None:
                                                              dataset=dataset,
                                                              model_name='GraphPooling')
 
-    # ----- Pseudotime -----
+    # ----- niche cluster -----
+    if consolidate_s_array is not None:
+        ori_data_df = load_original_data(options=options)
+        graph_pooling_output(ori_data_df=ori_data_df,
+                             dataset=dataset,
+                             rel_params=get_rel_params(
+                                 options=options, params=read_yaml_file(f'{options.preprocessing_dir}/samples.yaml')),
+                             consolidate_s_array=consolidate_s_array,
+                             output_dir=options.GNN_dir)
+
+    # ----- NT score -----
     if consolidate_s_array is not None and consolidate_out_adj_array is not None:
         NTScore(options=options,
                 dataset=dataset,